solved all warnings in my minigame class(the float warning again), however there is a weird error in the scene2.h when i create a minigame object. Might be a visual studio bug

+++ MyGraphics/Source/scene2_minigame.cpp
@@ -39,39 +39,39 @@ void minigame::init()
 }
 float minigame::piece1()
 {
-    piece1_rotate = 90 * piece1_state;
+    piece1_rotate = static_cast<float>(90 * piece1_state);
 }
 float minigame::piece2()
 {
-    piece2_rotate = 90 * piece2_state;
+    piece2_rotate = static_cast<float>(90 * piece2_state);
 }
 float minigame::piece3()
 {
-    piece3_rotate = 90 * piece3_state;
+    piece3_rotate = static_cast<float>(90 * piece3_state);
 }
 float minigame::piece4()
 {
-    piece4_rotate = 90 * piece4_state;
+    piece4_rotate = static_cast<float>(90 * piece4_state);
 }
 float minigame::piece5()
 {
-    piece5_rotate = 90 * piece5_state;
+    piece5_rotate = static_cast<float>(90 * piece5_state);
 }
 float minigame::piece6()
 {
-    piece6_rotate = 90 * piece6_state;
+    piece6_rotate = static_cast<float>(90 * piece6_state);
 }
 float minigame::piece7()
 {
-    piece7_rotate = 90 * piece7_state;
+    piece7_rotate = static_cast<float>(90 * piece7_state);
 }
 float minigame::piece8()
 {
-    piece8_rotate = 90 * piece8_state;
+    piece8_rotate = static_cast<float>(90 * piece8_state);
 }
 float minigame::piece9()
 {
-    piece9_rotate = 90 * piece9_state;
+    piece9_rotate = static_cast<float>(90 * piece9_state);
 }
 void minigame::updatePieceState()
 {
